encode-spoonOS-main/wallet-guardian/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
spoon-ai-sdk>=0.1.0
pytest>=8.3.0

